refactor: rename home page directory
