+++ lib/main.dart
@@ -1,18 +1,22 @@
 import "package:flutter/material.dart";
+import 'package:flutter/services.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:get/get.dart';
 import 'package:ilkda_client/view/12_read_book_page/read_book_page.dart';
 import 'package:ilkda_client/view/13_book_report_page/book_report_page.dart';
+import 'package:ilkda_client/view/20_book_shelf_page/book_shelf_page.dart';
+import 'package:ilkda_client/view_model/book_shelf_page_viewcontroller.dart';
 import 'package:kakao_flutter_sdk_user/kakao_flutter_sdk_user.dart';
 
+import 'constants/key.dart';
 import 'view/0_login_page/login_page.dart';
-import 'view/1_home_page/home_page.dart';
+import 'view/10_home_page/home_page.dart';
 import 'view_model/home_page_viewcontroller.dart';
 import 'view_model/user_controller.dart';
 import 'view/11_search_page/search_page.dart';
 
 void main() {
-  KakaoSdk.init(nativeAppKey: '4986af7734e2429d4f9cf6fe53bd864f');
+  KakaoSdk.init(nativeAppKey: '${kakaoNativeAppKey}');
   runApp(IlkdaClient());
 }
 
@@ -21,17 +25,18 @@ class IlkdaClient extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    SystemChrome.setEnabledSystemUIMode(SystemUiMode.immersive);
     return ScreenUtilInit(
       designSize: Size(360, 800),
       builder: (context, child) {
         return GetMaterialApp(
           debugShowCheckedModeBanner: false,
           title: '읽다',
-          initialRoute: "/LogIn",
+          initialRoute: "/",
           initialBinding: BindingsBuilder(() {}),
           getPages: [
             GetPage(
-              name: "/LogIn",
+              name: "/",
               page: () => const LogInPage(),
               binding: BindingsBuilder(() {
                 Get.put(UserController());
@@ -44,6 +49,7 @@ class IlkdaClient extends StatelessWidget {
               binding: BindingsBuilder(() {
                 Get.put(HomePageViewController());
               }),
+              transition: Transition.leftToRight,
             ),
             GetPage(
               name: "/Home/Search",
@@ -62,6 +68,16 @@ class IlkdaClient extends StatelessWidget {
               binding: BindingsBuilder(() {
                 Get.put(HomePageViewController());
               }),
+            ),
+
+            GetPage(
+              name: "/BookShelf",
+              page: () => const BookShelfPage(),
+              binding: BindingsBuilder(() {
+                Get.put(HomePageViewController());
+                Get.put(BookShelfPageViewController());
+              }),
+              transition: Transition.rightToLeft,
             ),
           ],
